with_python_sdk/pyproject.toml
[project]
name = "with-python-sdk"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.21.1",
]
